requirements.txt
numpy>=1.24
Pillow>=10.0
scikit-image>=0.21
scipy>=1.11
matplotlib>=3.7
PyYAML>=6.0
